VectorSpaceLeastSquares.computeGlobalIndex — Method

computeGlobalIndex(p::PiecewiseConstantBasis, x::AbstractVector{<:Real})

Compute the linear index of grid cell in which x lies. Return -1 if x is not inside the grid.

VectorSpaceLeastSquares.computeMeanF — Method

computeMeanF(x::AbstractVector{<:AbstractVector{T}}, f::Function)

Compute E[f(X)] where f: R → R is applied component-wise. Each entry of x is supposed to be a sample from the distribution of X.

VectorSpaceLeastSquares.computePolynomialTensor — Method

computePolynomialTensor(nVariates::Integer, degree::Integer)

Compute the full tensor representation of a multivariate polynomial

VectorSpaceLeastSquares.dtchebychev1d — Method

Recursive computation of the first derivative of the Tchebychev polynomials of any order.

x the evaluation point

n the order of the polynomial to be evaluated

n0 the rank of initialization

f_n the derivative of the polynomial of order n0.

f_n_1 the derivative of the polynomial of order n0 - 1

VectorSpaceLeastSquares.dtchebychev1d — Method

First derivative of the Tchebytchev polynomials

x the evaluation point

n the index of the polynomial whose first derivative is to be evaluated

VectorSpaceLeastSquares.getTx — Method

getTx(vslsq::VSLeastSquares{Tb, Tt, Td}) where {Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

Return the vector internally used to store the transformed data. Internal use only.

VectorSpaceLeastSquares.hermite1d — Method

Recursive function to compute Hermite polynomials of any order.

x evaluation point

n` the order of the polynomial to be evaluated

n0 the rank of the initialization

f_n0 used to store the polynomial of order n0

f_n0_1 used to store the polynomial of order n0 - 1

VectorSpaceLeastSquares.tchebychev1d — Method

Recursive function to compute Tchebychev polynomials of any order.

x evaluation point

n` the order of the polynomial to be evaluated

n0 the rank of the initialization

f_n0 used to store the polynomial of order n0

f_n0_1 used to store the polynomial of order n0 - 1

VectorSpaceLeastSquares.tchebychev1d — Method

Tchebychev polynomials of any order

x the evaluation point

n the order of the polynomial to be evaluated