Kernel Bases

It is possible to use a kernel basis instead of a basis of functions of a single variate (be it multi-dimensional): the functions are defined by kernels $K(x_i, \cdot) : \mathbb{R}^d \to \mathbb{R}$, where the nodes $x_i \in \mathbb{R}^d$ are defined as the training values

VectorSpaceLeastSquares.KernelBasis — Type

KernelBasis{Tk, Td} <: AbstractBasis where {Tk <: AbstractKernel, Td <: Real}

Represent a kernel basis where Tk must be a concrete implementation of an AbstractKernel.

VectorSpaceLeastSquares.KernelBasis — Method

KernelBasis(typeTd::Type{<:Real}, kernel::AbstractKernel, nVariates::Integer)

Kernel type

To create a KernelBasis, you must define the kernel to use, which is done by deriving a concrete type from AbstractKernel

VectorSpaceLeastSquares.AbstractKernel — Type

AbstractKernel

Abstract kernel structure. A concrete kernel structure must implement kernel and dkernel.

VectorSpaceLeastSquares.kernel — Method

kernel(k::AbstractKernel, node::AbstractVector{<:Real}, x::AbstractVector{<:Real})

Compute the value at x of the kernel centered at node.

VectorSpaceLeastSquares.dkernel — Method

dkernel(k::AbstractKernel, node::AbstractVector{<:Real}, x::AbstractVector{<:Real}, derivativeIndex::Integer)

Compute the first derivative w.r.t the derivativeIndex coordinate at x of the kernel centered at node.

Gaussian kernel

The Gaussian kernel is defined by $K(x,y) = \frac{1}{(2 \pi \sigma^2)^{d/2}} e^{-\frac{|x - y|^2}{2 \sigma^2}}.$

VectorSpaceLeastSquares.GaussianKernel — Type

GaussianKernel

Define Gaussian kernel

VectorSpaceLeastSquares.GaussianKernel — Method

GaussianKernel(sigmaSq::Td) where {Td<:Real}

Create a Gaussian kernel with variance sigmaSq.

VectorSpaceLeastSquares.kernel — Method

kernel(k::GaussianKernel{Td}, node::AbstractVector{Td}, x::AbstractVector{Td}) where {Td<:Real}

Compute the value at x of the Gaussian kernel centered at node.

VectorSpaceLeastSquares.dkernel — Method

dkernel(k::GaussianKernel{Td}, node::AbstractVector{Td}, x::AbstractVector{Td}, derivativeIndex::Integer) where {Td<:Real}

Compute the first derivative w.r.t the derivativeIndex coordinate at x of the Gaussian kernel centered at node.

VectorSpaceLeastSquares.d2kernel — Method

d2kernel(k::GaussianKernel{Td}, node::AbstractVector{Td}, x::AbstractVector{Td}, derivativeIndex1::Integer, derivativeIndex2::Integer) where {Td<:Real}

Compute the second derivative w.r.t the (derivativeIndex1,derivativeIndex2) coordinates at x of the Gaussian kernel centered at node.