VectorSpaceLeastSquares.AbstractBasis — Type

AbstractBasis

Super type for all bases. A basis must implement the following methods: nVariates, length, getType, isDifferentiable, value. If the basis contains only differentiable functions, in which case isDifferentiable returns true, it must also implement derivative. If the basis contains only twice differentiable functions, in which case isTwiceDifferentiable returns true, it must also implement secondDerivative.

VectorSpaceLeastSquares.AbstractKernel — Type

AbstractKernel

Abstract kernel structure. A concrete kernel structure must implement kernel and dkernel.

VectorSpaceLeastSquares.AbstractTransformation — Type

AbstractTransformation

Super type for all transformations.

A transformation is a function $\varphi: \mathbb{R}^d \to \mathbb{R}^d$, which is applied on the fly to the data before proceeding with the least squares problem. A transformation must implement apply! and jacobian.

To define a new transformation, define a new concrete subtype of AbstractTransformation and implement the corresponding apply! and jacobian methods. For instance the scaled log-transformation $\varphi(x) = \alpha \log(x)$ where $\alpha \in \mathbb{R}$ is defined as follows

struct LogTransformation <: AbstractTransformation
    scale::Vector{<:Real}
end

function apply!(t::LogTransformation, tx::AbstractVector{Td}, x::AbstractVector{Td}) where Td<:Real
    tx .= t.scale .* log.(x)
end

function jacobian(t::LogTransformation, x::AbstractVector{<:Real}, i::Integer, j::Integer)
    if i != j
        return 0.
    end
    t.scale[i] / x[i]
end

VectorSpaceLeastSquares.GaussianKernel — Method

GaussianKernel(sigmaSq::Td) where {Td<:Real}

Create a Gaussian kernel with variance sigmaSq.

VectorSpaceLeastSquares.GaussianKernel — Type

GaussianKernel

Define Gaussian kernel

VectorSpaceLeastSquares.GaussianTransformation — Method

GaussianTransformation(x::AbstractVector{<:AbstractVector{T}}) where T<:Real

Create a Gaussian transformation by setting α as the empirical mean and σ as the empirical standard deviation. Each entry of x is supposed to be one sample of the data.

VectorSpaceLeastSquares.GaussianTransformation — Type

GaussianTransformation{Td} <: AbstractTransformation where Td<:Real

Implement an Gaussian transformation of the data defined by $\varphi(x) = N((x - \alpha) / \sigma)$ where $N$ is the cdf of the standard Gaussian distribution

VectorSpaceLeastSquares.KernelBasis — Method

KernelBasis(typeTd::Type{<:Real}, kernel::AbstractKernel, nVariates::Integer)

VectorSpaceLeastSquares.KernelBasis — Type

KernelBasis{Tk, Td} <: AbstractBasis where {Tk <: AbstractKernel, Td <: Real}

Represent a kernel basis where Tk must be a concrete implementation of an AbstractKernel.

VectorSpaceLeastSquares.LinearTransformation — Method

LinearTransformation(x::AbstractVector{<:AbstractVector{T}}) where T<:Real

Create a linear transformation by setting α as the empirical mean and σ as the inverse of the empirical standard deviation. Each entry of x is supposed to be one sample of the data.

VectorSpaceLeastSquares.LinearTransformation — Type

LinearTransformation{Td} <: AbstractTransformation where Td<:Real

Implement a linear transformation of the data defined by $\varphi(x) = (x - \alpha) * \sigma$

VectorSpaceLeastSquares.LogNormalTransformation — Method

LogNormalTransformation(x::AbstractVector{<:AbstractVector{T}}) where T<:Real

Create a Log-normal transformation by setting α and σ as the empirical mean and variance of log(x) . Each entry of x is supposed to be one sample of the data.

VectorSpaceLeastSquares.LogNormalTransformation — Type

LogNormalTransformation{Td} <: AbstractTransformation where Td<:Real

Implement a Log-normal transformation of the data defined by $\varphi(x) = N((\log(x) - \alpha) / \sigma)$ where $N$ is the cdf of the standard Gaussian distribution.

VectorSpaceLeastSquares.PiecewiseConstantBasis — Method

PiecewiseConstantBasis(nVariates::Integer, nIntervals::Integer)

Create a PiecewiseConstantBasis with nIntervals along each direction.

VectorSpaceLeastSquares.PiecewiseConstantBasis — Method

PiecewiseConstantBasis(nVariates::Integer, nIntervals::Vector{<:Integer})

Create a PiecewiseConstantBasis by specifying the number of intervals per direction (possibly different to allow non squared grids).

VectorSpaceLeastSquares.PiecewiseConstantBasis — Type

PiecewiseConstantBasis

Represent a basis of local functions defined on $[0,1]^d$

nVariates is the dimension d of the space.

nIntervals is a vector of size d defining the number of sub-intervals to use along each dimension

VectorSpaceLeastSquares.PolynomialBasis — Method

PolynomialBasis(type::PolynomialType, nVariates::Integer, degree::Integer)

Create a polynomial basis with type, nVariates variables and total maximum degree

VectorSpaceLeastSquares.PolynomialBasis — Type

PolynomialBasis

Represent a multivariate polynomial

degree::Int64: maximum total degree

nVariates::Int6: number of variates

dim::Int6: dimension of the generated vector space

type::PolynomialType: a value from PolynomialType

tensor::SparseMatrixCSC{Int64, Int64}: the sparse tensor representation of the polynomial. Polynomials are stored by column.

VectorSpaceLeastSquares.PolynomialType — Type

PolynomialType

List the families of polynomials available through the PolynomialBasis type

Canonic

Hermite

Tchebychev

VectorSpaceLeastSquares.VSLeastSquares — Method

VSLeastSquares(basis::Tb, transform::Tt=VoidTransformation(), Td::Type=Float64) where {Tb<:AbstractBasis, Tt<:AbstractTransformation}

Create a VSLeastSquares object from basis and transform and capable of handling Td typed data.

VectorSpaceLeastSquares.VSLeastSquares — Type

VSLeastSquares{Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

The main object to solve a least squares problem.

VectorSpaceLeastSquares.VoidTransformation — Type

VoidTransformation <: AbstractTransformation

This transformation does nothing.

Base.length — Method

length(b::AbstractBasis)

Return the number of elements in the basis.

Base.length — Method

length(vslsq::VSLeastSquares{Tb, Tt, Td}) where {Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

Return the number of functions of the basis used to solve the least squares problem

Base.size — Method

size(b::AbstractBasis)

Return the tuple (nVariates, length).

Base.size — Method

size(vslsq::VSLeastSquares{Tb, Tt, Td}) where {Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

Return the tuple (nVariates, length)

VectorSpaceLeastSquares.apply! — Method

apply!(t::AbstractTransformation, tx::AbstractVector{Td}, x::AbstractVector{Td}) where Td<:Real

Apply the transformation t to x and store the result in tx.

VectorSpaceLeastSquares.d2kernel — Method

d2kernel(k::GaussianKernel{Td}, node::AbstractVector{Td}, x::AbstractVector{Td}, derivativeIndex1::Integer, derivativeIndex2::Integer) where {Td<:Real}

Compute the second derivative w.r.t the (derivativeIndex1,derivativeIndex2) coordinates at x of the Gaussian kernel centered at node.

VectorSpaceLeastSquares.derivative — Method

derivative(b::AbstractBasis, x::AbstractVector{<:Real}, index::Integer, i::Integer)

Compute the value of the first derivative of the index-th basis function w.r.t to the i variate at point x. Only available if isDifferentiable returns true.

VectorSpaceLeastSquares.derivative — Method

derivative(polType::PolynomialType, degree::Integer, x::Real)

Evaluate the first derivative of a 1d polynomial of type polType and degree degree at x

VectorSpaceLeastSquares.derivative — Method

derivative(vslsq::VSLeastSquares{Tb, LinearTransformation, Td}, x::AbstractVector{Td}, index::Integer) where {Tb<:AbstractBasis, Td<:Real}

Compute the partial derivative of the prediction w.r.t to the index variable for a linear transformation.

The method fit must have been called before.

VectorSpaceLeastSquares.derivative — Method

derivative(vslsq::VSLeastSquares{Tb, VoidTransformation, Td}, x::AbstractVector{Td}, index::Integer) where {Tb<:AbstractBasis, Td<:Real}

Compute the partial derivative of the prediction w.r.t to the index variable for a VoidTransformation.

The method fit must have been called before.

VectorSpaceLeastSquares.derivative — Method

derivative(b::KernelBasis{<:AbstractKernel, <:Real}, x::AbstractVector{<:Real}, index::Integer, derivativeIndex::Integer)

Compute the value of the first derivative of the kernel centered around the index-th node w.r.t to the derivativeIndex coordinate at point x.

VectorSpaceLeastSquares.derivative — Method

derivative(vslsq::VSLeastSquares{Tb, Tt, Td}, x::AbstractVector{Td}, index::Integer) where {Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

Compute the partial derivative of the prediction w.r.t to the index variable.

The method fit must have been called before.

VectorSpaceLeastSquares.derivative — Method

derivative(p::PolynomialBasis, x::AbstractVector{Td}, polIndex::Ti, derivativeIndex::Ti) where {Td<:Real, Ti<:Integer}

Evaluate the first partial derivative w.r.t variable derivativeIndex of the polIndex-th member of the polynomial basis p

VectorSpaceLeastSquares.dkernel — Method

dkernel(k::AbstractKernel, node::AbstractVector{<:Real}, x::AbstractVector{<:Real}, derivativeIndex::Integer)

Compute the first derivative w.r.t the derivativeIndex coordinate at x of the kernel centered at node.

VectorSpaceLeastSquares.dkernel — Method

dkernel(k::GaussianKernel{Td}, node::AbstractVector{Td}, x::AbstractVector{Td}, derivativeIndex::Integer) where {Td<:Real}

Compute the first derivative w.r.t the derivativeIndex coordinate at x of the Gaussian kernel centered at node.

VectorSpaceLeastSquares.fit — Method

fit(vslsq::VSLeastSquares{Tb, Tt, Td}, x::AbstractVector{<:AbstractVector{Td}}, y::AbstractVector{Td}, lambda::Td = Td(0)) where {Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

Solve the least squares problem.

VectorSpaceLeastSquares.fit — Method

fit(vslsq::VSLeastSquares{Tb, Tt, Td}, x::AbstractVector{<:AbstractVector{Td}}, y::AbstractVector{Td}, lambda::Td = Td(0)) where {Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

Solve the least squares problem.

VectorSpaceLeastSquares.fit — Method

fit(vslsq::VSLeastSquares{PiecewiseConstantBasis, Tt, Td}, x::AbstractVector{<:AbstractVector{Td}}, y::AbstractVector{Td}, lambda::Td = Td(0)) where {Tt<:AbstractTransformation, Td<:Real}

Solve the least squares problem using the specific structure of the PiecewiseConstantBasis.

VectorSpaceLeastSquares.getBasis — Method

getBasis(vslsq::VSLeastSquares{Tb, Tt, Td}) where {Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

Return the basis used to solve the least squares problem.

VectorSpaceLeastSquares.getCenter — Method

getCenter(t::LinearTransformation{<:Real})

Return the center α of the linear transformation

VectorSpaceLeastSquares.getCoefficients — Method

getCoefficients(vslsq::VSLeastSquares{Tb, Tt, Td}) where {Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

Return the coefficients solution to the least squares problem.

VectorSpaceLeastSquares.getMean — Method

getMean(t::LinearTransformation{<:Real})

Return the mean α of the Gaussian distribution

VectorSpaceLeastSquares.getMean — Method

getMean(t::LinearTransformation{<:Real})

Return the mean α of the underlying Gaussian distribution

VectorSpaceLeastSquares.getScale — Method

getScale(t::LinearTransformation{<:Real}) = t.scale

Return the scale σ of the linear transformation

VectorSpaceLeastSquares.getSigma — Method

getScale(t::LinearTransformation{<:Real}) = t.scale

Return the standard deviation σ of the Gaussian distribution

VectorSpaceLeastSquares.getSigma — Method

getScale(t::LinearTransformation{<:Real}) = t.scale

Return the standard deviation σ of the underlying Gaussian distribution

VectorSpaceLeastSquares.getType — Method

getType(b::AbstractBasis)

Return the internal basis type.

VectorSpaceLeastSquares.gradient — Method

gradient(vslsq::VSLeastSquares{Tb, Tt, Td}, x::AbstractVector{Td}) where {Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

Compute the gradient of the prediction at x.

The method fit must have been called before.

VectorSpaceLeastSquares.hessian — Method

hessian(vslsq::VSLeastSquares{Tb, Tt, Td}, x::AbstractVector{Td}) where {Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

Compute the hessian of the prediction at x. It only works if the transformation Tt is a void or a linear transformation.

The method fit must have been called before.

VectorSpaceLeastSquares.isDifferentiable — Method

isDifferentiable(b::AbstractBasis)

Return true if the functions in the basis are differentiable. In this case, a specific method derivative must be implemented.

VectorSpaceLeastSquares.isTwiceDifferentiable — Method

isTwiceDifferentiable(b::AbstractBasis)

Return true if the functions in the basis are twice differentiable. In this case, a specific method secondDerivative must be implemented.

VectorSpaceLeastSquares.jacobian — Method

jacobian(t::AbstractTransformation, x::AbstractVector{<:Real}, i::Integer, j::Integer)

Compute $\partial_{x_j} \varphi_i(x)$.

VectorSpaceLeastSquares.kernel — Method

kernel(k::AbstractKernel, node::AbstractVector{<:Real}, x::AbstractVector{<:Real})

Compute the value at x of the kernel centered at node.

VectorSpaceLeastSquares.kernel — Method

kernel(k::GaussianKernel{Td}, node::AbstractVector{Td}, x::AbstractVector{Td}) where {Td<:Real}

Compute the value at x of the Gaussian kernel centered at node.

VectorSpaceLeastSquares.nVariates — Method

nVariates(b::AbstractBasis)

Return the number of variates of the functions inside the basis.

VectorSpaceLeastSquares.nVariates — Method

nVariates(vslsq::VSLeastSquares{Tb, Tt, Td}) where {Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

Return the number of variates in the least squares problem

VectorSpaceLeastSquares.predict — Method

predict(vslsq::VSLeastSquares{Tb, Tt, Td}, x::AbstractVector{Td}) where {Tb<:AbstractBasis, Tt<:AbstractTransformation, Td<:Real}

Compute the value predicted by the least squares problem.

The method fit must have been called before.

VectorSpaceLeastSquares.predict — Method

predict(vslsq::VSLeastSquares{PiecewiseConstantBasis, Tt, Td}, x::AbstractVector{Td}) where {Tt<:AbstractTransformation, Td<:Real}

Compute the value predicted by the least squares problem using the specific structure of the PiecewiseConstantBasis.

The method fit must have been called before.

VectorSpaceLeastSquares.secondDerivative — Method

secondDerivative(b::AbstractBasis, x::AbstractVector{<:Real}, index::Integer, i1::Integer, i2::Integer)

Compute the value of the second derivative of the index-th basis function w.r.t to the i1, i2 variates at point x. Only available if isTwiceDifferentiable returns true.

VectorSpaceLeastSquares.secondDerivative — Method

secondDerivative(b::KernelBasis{<:AbstractKernel, <:Real}, x::AbstractVector{<:Real}, index::Integer, i1::Integer, i2::Integer)

Compute the value of the second derivative of the index-th basis function w.r.t to the i1, i2 variates at point x.

VectorSpaceLeastSquares.secondDerivative — Method

secondDerivative(polType::PolynomialType, degree::Integer, x::Real)

Evaluate the second derivative of a 1d polynomial of type polType and degree degree at x

VectorSpaceLeastSquares.secondDerivative — Method

secondDerivative(p::PolynomialBasis, x::AbstractVector{Td}, polIndex::Ti, i1::Ti, i2::Ti) where {Td<:Real, Ti<:Integer}

Evaluate the second partial derivative w.r.t variables i1, i2 of the polIndex-th member of the polynomial basis p

VectorSpaceLeastSquares.secondDerivative — Method

secondDerivative(vslsq::VSLeastSquares{Tb, LinearTransformation, Td}, x::AbstractVector{Td}, i::Integer, j::Integer) where {Tb<:AbstractBasis, Td<:Real}

Compute the second order partial derivative of the prediction w.r.t to the (i,j) variables for a linear transformation.

The method fit must have been called before.

VectorSpaceLeastSquares.secondDerivative — Method

secondDerivative(vslsq::VSLeastSquares{Tb, VoidTransformation, Td}, x::AbstractVector{Td}, i::Integer, j::Integer) where {Tb<:AbstractBasis, Td<:Real}

Compute the second order partial derivative of the prediction w.r.t to the (i,j) variables for a void transformation.

The method fit must have been called before.

VectorSpaceLeastSquares.value — Method

value(b::AbstractBasis, x::AbstractVector{<:Real}, index::Integer)

Compute the value of the index-th basis function at point x.

VectorSpaceLeastSquares.value — Method

value(p::PolynomialBasis, x::AbstractVector{<:Real}, i::Integer)

Evaluate the i-th function of a Polynomial basis p at point x

VectorSpaceLeastSquares.value — Method

value(polType::PolynomialType, degree::Integer, x::Real)

Evaluate a 1d polynomial of type polType and degree degree at x

VectorSpaceLeastSquares.value — Method

value(b::KernelBasis{<:AbstractKernel, <:Real}, x::AbstractVector{<:Real}, index::Integer)

Compute the value of the kernel centered around the index-th node at point x.