Vector Space bases

The vector space $\mathcal{V}$ is represented by the abstract type AbstractBasis.

VectorSpaceLeastSquares.AbstractBasis — Type

AbstractBasis

Super type for all bases. A basis must implement the following methods: nVariates, length, getType, isDifferentiable, value. If the basis contains only differentiable functions, in which case isDifferentiable returns true, it must also implement derivative.

VectorSpaceLeastSquares.nVariates — Method

nVariates(b::AbstractBasis)

Return the number of variates of the functions inside the basis.

Base.length — Method

length(b::AbstractBasis)

Return the number of elements in the basis.

Base.size — Method

size(b::AbstractBasis)

Return the tuple (nVariates, length).

VectorSpaceLeastSquares.getType — Method

getType(b::AbstractBasis)

Return the internal basis type.

VectorSpaceLeastSquares.isDifferentiable — Method

isDifferentiable(b::AbstractBasis)

Return true if the functions in the basis are differentiable. In this case, a specific method derivative must be implemented.

VectorSpaceLeastSquares.value — Method

value(b::AbstractBasis, x::AbstractVector{<:Real}, index::Integer)

Compute the value of the index-th basis function at point x.

VectorSpaceLeastSquares.derivative — Method

derivative(b::AbstractBasis, x::AbstractVector{<:Real}, index::Integer, derivativeIndex::Integer)

Compute the value of the first derivative of the index-th basis function w.r.t to the derivativeIndex variate at point x.

Polynomial Bases

Polynomial Bases are implemented using the PolynomialBasis type

VectorSpaceLeastSquares.PolynomialBasis — Type

PolynomialBasis

Represent a multivariate polynomial

degree::Int64: maximum total degree

nVariates::Int6: number of variates

dim::Int6: dimension of the generated vector space

type::PolynomialType: a value from PolynomialType

tensor::SparseMatrixCSC{Int64, Int64}: the sparse tensor representation of the polynomial. Polynomials are stored by column.

The following families of polynomials are available

VectorSpaceLeastSquares.PolynomialType — Type

PolynomialType

List the families of polynomials available through the PolynomialBasis type

Canonic

Hermite

Tchebychev

A PolynomialBasis object can be created using

VectorSpaceLeastSquares.PolynomialBasis — Method

PolynomialBasis(type::PolynomialType, nVariates::Integer, degree::Integer)

Create a polynomial basis with type, nVariates variables and total maximum degree

Piecewise constant functions (local bases)

Piecewise constant functions can be efficiently obtained from a basis of local functions with disjoint supports. Such bases are implemented using the PiecewiseConstantBasis

VectorSpaceLeastSquares.PiecewiseConstantBasis — Type

PiecewiseConstantBasis

Represent a basis of local functions defined on $[0,1]^d$

nVariates is the dimension d of the space.

nIntervals is a vector of size d defining the number of sub-intervals to use along each dimension

A PiecewiseConstantBasis object can be created using

VectorSpaceLeastSquares.PiecewiseConstantBasis — Method

PiecewiseConstantBasis(nVariates::Integer, nIntervals::Integer)

Create a PiecewiseConstantBasis with nIntervals along each direction.

VectorSpaceLeastSquares.PiecewiseConstantBasis — Method

PiecewiseConstantBasis(nVariates::Integer, nIntervals::Vector{<:Integer})

Create a PiecewiseConstantBasis by specifying the number of intervals per direction (possibly different to allow non squared grids).