Transformations

VectorSpaceLeastSquares.AbstractTransformation — Type

AbstractTransformation

Super type for all transformations.

A transformation is a function $\varphi: \mathbb{R}^d \to \mathbb{R}^d$, which is applied on the fly to the data before proceeding with the least squares problem. A transformation must implement apply! and jacobian.

To define a new transformation, define a new concrete subtype of AbstractTransformation and implement the corresponding apply! and jacobian methods. For instance the scaled log-transformation $\varphi(x) = \alpha \log(x)$ where $\alpha \in \mathbb{R}$ is defined as follows

struct LogTransformation <: AbstractTransformation
    scale::Vector{<:Real}
end

function apply!(t::LogTransformation, tx::AbstractVector{Td}, x::AbstractVector{Td}) where Td<:Real
    tx .= t.scale .* log.(x)
end

function jacobian(t::LogTransformation, x::AbstractVector{<:Real}, i::Integer, j::Integer)
    if i != j
        return 0.
    end
    t.scale[i] / x[i]
end

The following methods are available and must be defined for any concrete subtype of AbstractTransformation

VectorSpaceLeastSquares.apply! — Function

apply!(t::AbstractTransformation, tx::AbstractVector{Td}, x::AbstractVector{Td}) where Td<:Real

Apply the transformation t to x and store the result in tx.

VectorSpaceLeastSquares.jacobian — Function

jacobian(t::AbstractTransformation, x::AbstractVector{<:Real}, i::Integer, j::Integer)

Compute $\partial_{x_j} \varphi_i(x)$.

Void transformation

It corresponds to $\varphi(x) = x$ for $x \in \mathbb{R}^d$.

VectorSpaceLeastSquares.VoidTransformation — Type

VoidTransformation <: AbstractTransformation

This transformation does nothing.

Linear transformation

It corresponds to $\varphi: \mathbb{R}^d \to \mathbb{R}^d$ such that $\varphi(x) = (x - \alpha) * \sigma$ for where $\alpha, \sigma \in \mathbb{R}^d$ and * denotes a term by term multiplication. In the following, $\alpha$ is called the center and $\sigma$ is called the scale

VectorSpaceLeastSquares.LinearTransformation — Type

LinearTransformation{Td} <: AbstractTransformation where Td<:Real

Implement a linear transformation of the data defined by $\varphi(x) = (x - \alpha) * \sigma$

VectorSpaceLeastSquares.getCenter — Function

getCenter(t::LinearTransformation{<:Real})

Return the center α of the linear transformation

VectorSpaceLeastSquares.getScale — Function

getScale(t::LinearTransformation{<:Real}) = t.scale

Return the scale σ of the linear transformation

VectorSpaceLeastSquares.LinearTransformation — Method

LinearTransformation(x::AbstractVector{<:AbstractVector{T}}) where T<:Real

Create a linear transformation by setting α as the empirical mean and σ as the inverse of the empirical standard deviation. Each entry of x is supposed to be one sample of the data.

Gaussian transformation

It corresponds to $\varphi: \mathbb{R}^d \to \mathbb{R}^d$ such that $\varphi(x) = \mathcal{N}((x - \alpha) / \sigma)$ where $\alpha, \sigma \in \mathbb{R}^d$, $\mathcal{N}$ is the cumulative distribution function of the standard normal distribution and / denotes a term by term multiplication.

VectorSpaceLeastSquares.GaussianTransformation — Type

GaussianTransformation{Td} <: AbstractTransformation where Td<:Real

Implement an Gaussian transformation of the data defined by $\varphi(x) = N((x - \alpha) / \sigma)$ where $N$ is the cdf of the standard Gaussian distribution

VectorSpaceLeastSquares.getMean — Method

getMean(t::LinearTransformation{<:Real})

Return the mean α of the Gaussian distribution

VectorSpaceLeastSquares.getSigma — Method

getScale(t::LinearTransformation{<:Real}) = t.scale

Return the standard deviation σ of the Gaussian distribution

VectorSpaceLeastSquares.GaussianTransformation — Method

GaussianTransformation(x::AbstractVector{<:AbstractVector{T}}) where T<:Real

Create a Gaussian transformation by setting α as the empirical mean and σ as the empirical standard deviation. Each entry of x is supposed to be one sample of the data.

Log-normal transformation

It corresponds to $\varphi: \mathbb{R}^d \to \mathbb{R}^d$ such that $\varphi(x) = \mathcal{N}((\log(x) - \alpha) / \sigma)$ where $\alpha, \sigma \in \mathbb{R}^d$, $\mathcal{N}$ is the cumulative distribution function of the standard normal distribution and / denotes a term by term multiplication.

VectorSpaceLeastSquares.LogNormalTransformation — Type

LogNormalTransformation{Td} <: AbstractTransformation where Td<:Real

Implement a Log-normal transformation of the data defined by $\varphi(x) = N((\log(x) - \alpha) / \sigma)$ where $N$ is the cdf of the standard Gaussian distribution.

VectorSpaceLeastSquares.getMean — Method

getMean(t::LinearTransformation{<:Real})

Return the mean α of the underlying Gaussian distribution

VectorSpaceLeastSquares.getSigma — Method

getScale(t::LinearTransformation{<:Real}) = t.scale

Return the standard deviation σ of the underlying Gaussian distribution

VectorSpaceLeastSquares.LogNormalTransformation — Method

LogNormalTransformation(x::AbstractVector{<:AbstractVector{T}}) where T<:Real

Create a Log-normal transformation by setting α and σ as the empirical mean and variance of log(x) . Each entry of x is supposed to be one sample of the data.